lightweight_real_esrgan_anime/requirements.txt
gradio
huggingface-hub
onnxruntime
Pillow
numpy